+++ crates/miden-protocol/src/account/account_id/v1/prefix.rs
@@ -5,7 +5,7 @@ use core::hash::Hash;
 use miden_core::Felt;
 
 use crate::account::account_id::v1::{self, validate_prefix};
-use crate::account::{AccountIdVersion, AccountStorageMode, AccountType};
+use crate::account::{AccountIdVersion, AccountType};
 use crate::errors::AccountIdError;
 use crate::utils::serde::{
     ByteReader,
@@ -77,29 +77,12 @@ impl AccountIdPrefixV1 {
     /// See [`AccountIdPrefix::account_type`](crate::account::AccountIdPrefix::account_type) for
     /// details.
     pub fn account_type(&self) -> AccountType {
-        v1::extract_type(self.prefix.as_canonical_u64())
-    }
-
-    /// See [`AccountIdPrefix::is_faucet`](crate::account::AccountIdPrefix::is_faucet) for details.
-    pub fn is_faucet(&self) -> bool {
-        self.account_type().is_faucet()
-    }
-
-    /// See [`AccountIdPrefix::is_regular_account`](crate::account::AccountIdPrefix::is_regular_account) for
-    /// details.
-    pub fn is_regular_account(&self) -> bool {
-        self.account_type().is_regular_account()
-    }
-
-    /// See [`AccountIdPrefix::storage_mode`](crate::account::AccountIdPrefix::storage_mode) for
-    /// details.
-    pub fn storage_mode(&self) -> AccountStorageMode {
-        v1::extract_storage_mode(self.prefix.as_canonical_u64())
+        v1::extract_account_type(self.prefix.as_canonical_u64())
     }
 
     /// See [`AccountIdPrefix::is_public`](crate::account::AccountIdPrefix::is_public) for details.
     pub fn is_public(&self) -> bool {
-        self.storage_mode().is_public()
+        self.account_type().is_public()
     }
 
     /// See [`AccountIdPrefix::version`](crate::account::AccountIdPrefix::version) for details.