+++ crates/miden-protocol/src/note/mod.rs
@@ -77,8 +77,8 @@ pub use file::NoteFile;
 /// into a recipient object.
 ///
 /// Note details can be reduced to a [NoteDetailsCommitment]. Together with the note metadata,
-/// this commitment determines the public [NoteId]. Full note details can also be reduced to a
-/// [Nullifier], which is known only to entities which have access to full note details.
+/// this commitment determines the public [NoteId]. Full note details and metadata can also be
+/// reduced to a [Nullifier], which is known only to entities which have access to full note data.
 ///
 /// Fungible and non-fungible asset transfers are done by moving assets to the note's assets. The
 /// note's script determines the conditions required for the note consumption, i.e. the target
@@ -122,7 +122,7 @@ impl Note {
         let details = NoteDetails::new(assets, recipient);
         let metadata = NoteMetadata::new(partial_metadata, &attachments);
         let header = NoteHeader::new(details.commitment(), metadata);
-        let nullifier = details.nullifier();
+        let nullifier = Nullifier::from_details_and_metadata(&details, &metadata);
 
         Self { header, details, attachments, nullifier }
     }